requirements-r2.txt
boto3==1.37.38
Pillow==11.2.1
pillow-heif==0.22.0
